move logging setup to testing
