+++ .travis.yml
@@ -19,7 +19,7 @@ before_install:
   - export BRIZO_VERSION=v0.5.0
   - export AQUARIUS_VERSION=v0.4.2
   - rm -rf "${HOME}/.ocean/keeper-contracts/artifacts"
-  - bash -x start_ocean.sh --no-brizo --no-pleuston --local-spree-node 2>&1 > start_ocean.log &
+  - bash -x start_ocean.sh --no-brizo --no-commons --no-faucet --no-events-handler --local-spree-node 2>&1 > start_ocean.log &
   - cd ..
   - |
       for i in $(seq 1 50); do
