+++ tests/resources/helper_functions.py
@@ -5,6 +5,11 @@ import json
 import os
 import pathlib
 import uuid
+import logging
+import logging.config
+import coloredlogs
+import yaml
+
 
 from ocean_keeper.utils import get_account
 from ocean_utils.ddo.ddo import DDO
@@ -117,3 +122,25 @@ def get_metadata():
     with open(path, 'r') as file_handle:
         metadata = file_handle.read()
     return json.loads(metadata)
+
+def setup_logging(default_path='logging.yaml', default_level=logging.INFO, env_key='LOG_CFG'):
+    """Logging setup."""
+    path = default_path
+    value = os.getenv(env_key, None)
+    if value:
+        path = value
+    if os.path.exists(path):
+        with open(path, 'rt') as file:
+            try:
+                config = yaml.safe_load(file.read())
+                logging.config.dictConfig(config)
+                coloredlogs.install()
+                logging.info(f'Logging configuration loaded from file: {path}')
+            except Exception as ex:
+                print(ex)
+                print('Error in Logging Configuration. Using default configs')
+                logging.basicConfig(level=default_level)
+                coloredlogs.install(level=default_level)
+    else:
+        logging.basicConfig(level=default_level)
+        coloredlogs.install(level=default_level)
